+++ test/plausible/event/system_events_test.exs
@@ -1,0 +1,5 @@
+defmodule Plausible.Event.SystemEventsTest do
+  use ExUnit.Case, async: true
+
+  doctest Plausible.Event.SystemEvents, import: true
+end