+++ CHANGELOG.md
@@ -36,6 +36,7 @@ All notable changes to this project will be documented in this file.
 - Fixed issue with site guests in Editor role and team members in Editor role not being able to change the domain of site
 - Fixed direct dashboard links that use legacy dashboard filters containing URL encoded special characters (e.g. character `ê` in the legacy filter `?page=%C3%AA`)
 - Fix bug with tracker script config cache that made requests for certain cached scripts give error 500
+- Fix issue with all non-interactive events being counted as interactive
 
 ## v3.1.0 - 2025-11-13